// frontend/src/app/core/generated/circabc/model/eventDefinition.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateInfo } from './dateInfo';
import { RepeatsInfo } from './repeatsInfo';
import { AttendantsInfo } from './attendantsInfo';


/**
 * definition of an event or meeting. Used to create and retrieve information about it
 */
export interface EventDefinition { 
    appointmentTypeEvent: boolean;
    id?: string;
    igId?: string;
    title: string;
    dateInfo: DateInfo;
    language: string;
    timezone: string;
    eventAbstract?: string;
    invitationMessage?: string;
    location?: string;
    repeatsInfo: RepeatsInfo;
    attendantsInfo: AttendantsInfo;
    enableNotification?: boolean;
    contactName: string;
    contactPhone: string;
    contactEmail: string;
    contactUrl?: string;
    meetingPublicAvailability: boolean;
    meetingOrganisation?: string;
    meetingAgenda?: string;
    meetingType: string;
    meetingLibrarySection?: string;
    eventType: EventDefinition.EventTypeEnum;
    eventPriority: EventDefinition.EventPriorityEnum;
}
export namespace EventDefinition {
    export const EventTypeEnum = {
        Task: 'Task',
        Appointment: 'Appointment',
        Other: 'Other'
    } as const;
    export type EventTypeEnum = typeof EventTypeEnum[keyof typeof EventTypeEnum];
    export const EventPriorityEnum = {
        Low: 'Low',
        Medium: 'Medium',
        High: 'High',
        Urgent: 'Urgent'
    } as const;
    export type EventPriorityEnum = typeof EventPriorityEnum[keyof typeof EventPriorityEnum];
}