// frontend/src/app/core/generated/circabc/model/repeatsInfo.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * in case an event/meeting repeats, stores this information
 */
export interface RepeatsInfo { 
    mainOccurence?: RepeatsInfo.MainOccurenceEnum;
    timesOccurence?: RepeatsInfo.TimesOccurenceEnum;
    everyTimesOccurence?: RepeatsInfo.EveryTimesOccurenceEnum;
    times?: number;
    everyTime?: number;
}
export namespace RepeatsInfo {
    export const MainOccurenceEnum = {
        OnlyOnce: 'OnlyOnce',
        Times: 'Times',
        EveryTimes: 'EveryTimes'
    } as const;
    export type MainOccurenceEnum = typeof MainOccurenceEnum[keyof typeof MainOccurenceEnum];
    export const TimesOccurenceEnum = {
        Daily: 'Daily',
        Weekly: 'Weekly',
        EveryTwoWeeks: 'EveryTwoWeeks',
        MonthlyByDate: 'MonthlyByDate',
        MonthlyByWeekday: 'MonthlyByWeekday',
        Yearly: 'Yearly',
        MondayToFriday: 'MondayToFriday',
        MondayWednseyFriday: 'MondayWednseyFriday',
        TuesdayThursday: 'TuesdayThursday'
    } as const;
    export type TimesOccurenceEnum = typeof TimesOccurenceEnum[keyof typeof TimesOccurenceEnum];
    export const EveryTimesOccurenceEnum = {
        Days: 'days',
        Weeks: 'weeks',
        Months: 'months'
    } as const;
    export type EveryTimesOccurenceEnum = typeof EveryTimesOccurenceEnum[keyof typeof EveryTimesOccurenceEnum];
}